+++ src/index.tsx
@@ -8,4 +8,4 @@ export * from "./SidebarDrawer";
 export * from "./SquareButton";
 export * from "./Alert";
 export * from "./Notification";
-export * from "./UserChip";
+export * from "./UserChipMenu";
